+++ lib/plugins/aws/package/compile/events/cloudWatchLog.js
