+++ lib/plugins/index.js
@@ -46,7 +46,7 @@ module.exports = [
   require('./aws/package/compile/events/alexaSmartHome.js'),
   require('./aws/package/compile/events/iot/index.js'),
   require('./aws/package/compile/events/cloudWatchEvent.js'),
-  require('./aws/package/compile/events/cloudWatchLog/index.js'),
+  require('./aws/package/compile/events/cloudWatchLog.js'),
   require('./aws/package/compile/events/cognitoUserPool/index.js'),
   require('./aws/package/compile/events/eventBridge/index.js'),
   require('./aws/package/compile/events/sqs/index.js'),
